@@ -4,18 +4,21 @@
   "private": true,
   "dependencies": {
     "axios": "^0.17.0",
+    "babel-plugin-emotion": "^8.0.9",
+    "emotion": "^8.0.9",
     "firebase": "^4.3.0",
     "moment": "^2.19.1",
     "prop-types": "^15.6.0",
     "react": "^15.6.1",
     "react-bootstrap": "^0.31.0",
     "react-dom": "^15.6.1",
+    "react-emotion": "^8.0.9",
     "react-router": "^4.1.1",
     "react-router-bootstrap": "^0.24.2",
     "react-router-dom": "^4.1.1",
-    "uuid": "^3.1.0",
     "react-stripe-elements": "^1.2.0",
-    "styled-components": "^2.1.0"
+    "styled-components": "^2.1.0",
+    "uuid": "^3.1.0"
   },
   "devDependencies": {
     "autoprefixer": "7.1.0",
@@ -32,13 +35,13 @@
     "enzyme": "^3.1.0",
     "enzyme-adapter-react-15": "^1.0.1",
     "enzyme-to-json": "^3.2.0",
-    "eslint": "3.19.0",
-    "eslint-config-react-app": "^1.0.4",
-    "eslint-loader": "1.7.1",
+    "eslint": "^4.10.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-loader": "^1.9.0",
     "eslint-plugin-flowtype": "2.33.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "5.0.3",
-    "eslint-plugin-react": "7.0.1",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-react": "^7.4.0",
     "extract-text-webpack-plugin": "2.1.0",
     "file-loader": "0.11.1",
     "fs-extra": "3.0.1",
