@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Routes } from './Routes';
+import Routes from './Routes';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
