@@ -2,22 +2,24 @@ import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import { Home } from './containers/Home';
-import { Volunteers } from './pages/Volunteers';
-import { Students } from './pages/Students';
-import Event from './pages/Event';
-import { CreateEvent } from './pages/CreateEvent';
-
-import { Events } from './pages/Events';
-import { Partners } from './pages/Partners';
-import { About } from './pages/About';
-import Apply from './pages/Apply';
-import { MeetTheTeam } from './pages/MeetTheTeam';
-import { Donate } from './pages/Donate';
+import {
+  Event,
+  ApplySuccess,
+  Apply,
+  Volunteers,
+  Students,
+  CreateEvent,
+  Events,
+  Partners,
+  About,
+  MeetTheTeam,
+  Donate,
+} from './pages';
 
 import Navigation from './components/navigation';
 import Footer from './components/footer';
 
-export const Routes = () => (
+const Routes = () => (
   <Router>
     <div>
       <Navigation />
@@ -28,7 +30,8 @@ export const Routes = () => (
       <Route exact path="/event/create" component={CreateEvent} />
       <Route exact path="/students" component={Students} />
       <Route exact path="/volunteers" component={Volunteers} />
-      <Route exact path="/apply/:form_type" component={Apply} />
+      <Route exact path="/apply/:formType" component={Apply} />
+      <Route exact path="/apply/success/:formType" component={ApplySuccess} />
       <Route exact path="/partners" component={Partners} />
       <Route exact path="/meet-the-team" component={MeetTheTeam} />
       <Route exact path="/donate" component={Donate} />
@@ -36,3 +39,5 @@ export const Routes = () => (
     </div>
   </Router>
 );
+
+export default Routes;
