@@ -3,7 +3,7 @@ const mediaQueries = {
   sm: '576px',
   md: '768px',
   lg: '992px',
-  xl: '1200px'
+  xl: '1200px',
 };
 
 export default mediaQueries;
